# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Library API Documentation",
        "description": "API documentation for the Library Management System",
        "contact": {
            "email": "admin@library.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Library API Server"
        }
    ],
    "paths": {
...
        "/api/borrowings": {
            "post": {
                "tags": [
                    "Borrowings"
                ],
                "summary": "Borrow a book",
                "operationId": "45b955f2786ecaa9a50fafe834c26674",
                "parameters": [
                    {
                        "name": "book_id",
                        "in": "query",
                        "description": "Book ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book borrowed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "borrowing": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Book not found"
                    },
                    "403": {
                        "description": "Unauthorized to borrow this book"
                    }
                },
                "security": {
                    "bearerAuth": {}
                }
            }
        },
        "/api/borrowing-return/{id}": {
            "post": {
                "tags": [
                    "Borrowings"
                ],
                "summary": "Return a borrowed book",
                "operationId": "b1633181e38f520b298898d50f6c4bc8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Borrowing ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book returned successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "borrowing": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Borrowing record not found"
                    },
                    "403": {
                        "description": "Unauthorized to return this book"
                    }
                },
                "security": {
                    "bearerAuth": {}
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}